requirements.txt
beautifulsoup4
opencv-python
numpy
Pillow
imutils
imagehash
requests